+++ rest/src/decorators/index.ts
@@ -1,2 +1,3 @@
 export * from './handle404.decorator';
 export * from './ignoreClientErrors.decorator';
+export * from './handle400WithValidations.decorator';