+++ rest/src/index.ts
@@ -2,4 +2,4 @@ export * from './decorators';
 export * from './middlewares';
 
 export * from './misc/restHttpError.interface';
-
+export * from './misc/withRestClientContext';